Added config to spring-data and create a simple repo

+++ api/src/main/java/br/com/thiagoaos/testitau/exception/TweetNotFoundException.java
@@ -1,7 +1,7 @@
 package br.com.thiagoaos.testitau.exception;
 
 public class TweetNotFoundException extends RuntimeException {
-    public TweetNotFoundException(Integer id) {
+    public TweetNotFoundException(long id) {
         super("Tweet with id not found " + id);
     }
 }
